Build as static library and minor refactoring
