+++ rviz_visual_tests/CMakeLists.txt
@@ -15,6 +15,7 @@ add_definitions(-D_BUILD_DIR_PATH="${CMAKE_CURRENT_BINARY_DIR}")
 add_definitions(-D_SRC_DIR_PATH="${CMAKE_CURRENT_SOURCE_DIR}")
 
 # Creates a directory where the test screenshots will be saved.
+file(REMOVE_RECURSE ${CMAKE_CURRENT_BINARY_DIR}/test_images)
 file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/test_images)
 
 # Decides whether the test or the reference screenshots will be taken.
@@ -47,57 +48,90 @@ find_package(Qt5 REQUIRED COMPONENTS Widgets Test)
 
 find_package(ament_cmake_gtest REQUIRED)
 
+set(visual_test_framework_source_files
+  tests/example_nodes.hpp
+  include/rviz_visual_tests/page_objects/base_page_object.hpp
+  src/page_objects/base_page_object.cpp
+  include/rviz_visual_tests/page_objects/camera_display_page_object.hpp
+  src/page_objects/camera_display_page_object.cpp
+  include/rviz_visual_tests/page_objects/grid_display_page_object.hpp
+  src/page_objects/grid_display_page_object.cpp
+  include/rviz_visual_tests/page_objects/image_display_page_object.hpp
+  src/page_objects/image_display_page_object.cpp
+  include/rviz_visual_tests/page_objects/page_object_with_window.hpp
+  src/page_objects/page_object_with_window.cpp
+  include/rviz_visual_tests/page_objects/point_cloud_display_page_object.hpp
+  src/page_objects/point_cloud_display_page_object.cpp
+  include/rviz_visual_tests/page_objects/polygon_display_page_object.hpp
+  src/page_objects/polygon_display_page_object.cpp
+  src/internal/rviz_scene_test.cpp
+  include/rviz_visual_tests/internal/rviz_scene_test.hpp
+  include/rviz_visual_tests/internal/display_handler.hpp
+  src/internal/display_handler.cpp
+  src/internal/image_tester.cpp
+  include/rviz_visual_tests/internal/image_tester.hpp
+  include/rviz_visual_tests/internal/test_helpers.hpp
+  src/internal/test_helpers.cpp
+  include/rviz_visual_tests/internal/executor.hpp
+  src/internal/executor.cpp
+  include/rviz_visual_tests/internal/visual_test.hpp
+  src/internal/visual_test.cpp
+  include/rviz_visual_tests/visual_test_fixture.hpp
+  src/visual_test_fixture.cpp)
+
+set(visual_tests_target_libaries
+  rviz_common::rviz_common
+  Qt5::Widgets
+  Qt5::Test)
+
+add_library(rviz_visual_tests STATIC ${visual_test_framework_source_files})
+
+target_include_directories(rviz_visual_tests PUBLIC
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
+
+target_link_libraries(rviz_visual_tests ${visual_tests_target_libaries} ${GTEST_LIBRARIES})
+
+# export information to downstream packages
+ament_export_interfaces(rviz_visual_tests)
+ament_export_dependencies(Qt5)
+ament_export_dependencies(rviz_common)
+
+install(
+  TARGETS rviz_visual_tests
+  EXPORT rviz_visual_tests
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+  INCLUDES DESTINATION include
+)
+
+install(
+  DIRECTORY include/
+  DESTINATION include
+)
+
 if(BUILD_TESTING)
-    # TODO(wjwwood): replace this with ament_lint_auto() and/or add the copyright linter back
-    find_package(ament_cmake_cppcheck REQUIRED)
-    find_package(ament_cmake_cpplint REQUIRED)
-    find_package(ament_cmake_lint_cmake REQUIRED)
-    find_package(ament_cmake_uncrustify REQUIRED)
-    ament_cppcheck()
-    ament_cpplint()
-    ament_lint_cmake()
-    ament_uncrustify()
-    if(EnableVisualTests STREQUAL "TRUE")
-      ament_add_gtest(
+  # TODO(wjwwood): replace this with ament_lint_auto() and/or add the copyright linter back
+  find_package(ament_cmake_cppcheck REQUIRED)
+  find_package(ament_cmake_cpplint REQUIRED)
+  find_package(ament_cmake_lint_cmake REQUIRED)
+  find_package(ament_cmake_uncrustify REQUIRED)
+  ament_cppcheck()
+  ament_cpplint()
+  ament_lint_cmake()
+  ament_uncrustify()
+
+  if(EnableVisualTests STREQUAL "TRUE")
+    ament_add_gtest(
+      example_test
+      tests/example_test.cpp
+      ${visual_test_framework_source_files})
+    if(TARGET example_test)
+      target_link_libraries(
         example_test
-        tests/example_test.cpp
-        tests/example_nodes.hpp
-        src/page_objects/base_page_object.hpp
-        src/page_objects/base_page_object.cpp
-        src/page_objects/camera_display_page_object.hpp
-        src/page_objects/camera_display_page_object.cpp
-        src/page_objects/image_display_page_object.hpp
-        src/page_objects/image_display_page_object.cpp
-        src/page_objects/page_object_with_window.hpp
-        src/page_objects/page_object_with_window.cpp
-        src/page_objects/point_cloud_display_page_object.hpp
-        src/page_objects/point_cloud_display_page_object.cpp
-        src/page_objects/grid_display_page_object.hpp
-        src/page_objects/grid_display_page_object.cpp
-        src/page_objects/polygon_display_page_object.hpp
-        src/page_objects/polygon_display_page_object.cpp
-        src/internal/rviz_scene_test.cpp
-        src/internal/rviz_scene_test.hpp
-        src/internal/display_handler.hpp
-        src/internal/display_handler.cpp
-        src/internal/image_tester.cpp
-        src/internal/image_tester.hpp
-        src/internal/test_helpers.hpp
-        src/internal/test_helpers.cpp
-        src/internal/executor.hpp
-        src/internal/executor.cpp
-        src/internal/visual_test.hpp
-        src/internal/visual_test.cpp
-        src/visual_test_fixture.hpp
-        src/visual_test_fixture.cpp)
-      if(TARGET example_test)
-        target_link_libraries(
-          example_test
-          rviz_common::rviz_common
-          Qt5::Widgets
-          Qt5::Test)
-      endif()
+        ${visual_tests_target_libaries})
     endif()
+  endif()
 endif()
 
 ament_package()
