+++ rviz_visual_tests/include/rviz_visual_tests/page_objects/camera_display_page_object.hpp
@@ -27,35 +27,37 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef PAGE_OBJECTS__POINT_CLOUD_DISPLAY_PAGE_OBJECT_HPP_
-#define PAGE_OBJECTS__POINT_CLOUD_DISPLAY_PAGE_OBJECT_HPP_
+#ifndef RVIZ_VISUAL_TESTS__PAGE_OBJECTS__CAMERA_DISPLAY_PAGE_OBJECT_HPP_
+#define RVIZ_VISUAL_TESTS__PAGE_OBJECTS__CAMERA_DISPLAY_PAGE_OBJECT_HPP_
 
 #include <memory>
 #include <vector>
 
-#include <QtWidgets>  // NOLINT
+#include "rviz_visual_tests/page_objects/page_object_with_window.hpp"
 
-#include "base_page_object.hpp"
-
-class PointCloudDisplayPageObject : public BasePageObject
+class CameraDisplayPageObject : public PageObjectWithWindow
 {
 public:
-  PointCloudDisplayPageObject(
+  CameraDisplayPageObject(
     int display_id,
     std::shared_ptr<Executor> executor,
     std::shared_ptr<std::vector<int>> all_displays_ids);
 
-  void setSizeMeters(QString meters_size);
-  void setSizePixels(QString pixels_size);
-  void setStyle(QString points_style);
-  void setSelectable(bool selectable);
-  void setAlpha(QString alpha);
-  void setDecayTime(QString decay_time);
-  void setQueueSize(QString queue_size);
+  /* @param display_name Name of the display whose visibility must be changed
+   * @param relative_row_index Sub-property index relative to the parent property (i.e. Visibility)
+   * @param display_visibility the value to set for the sub-property
+   */
+  void setDisplayVisibilityInRenderWindow(
+    QString display_name, int relative_row_index, bool display_visibility);
+  void setTopic(QString topic);
   void setUnreliable(bool unreliable);
-  void setPositionTransformer(QString position_transformer);
-  void setColorTransformer(QString color_transformer);
-  void setColor(int red, int green, int blue);
+  void setQueueSize(QString queue_size);
+  void setImageRendering(QString image_rendering);
+  void setOverlayAlpha(QString overlay_alpha);
+  void setZoomFacor(QString zoom_factor);
+  void setVisibility(bool visibility);
+
+  void setRenderWindow() override;
 };
 
-#endif  // PAGE_OBJECTS__POINT_CLOUD_DISPLAY_PAGE_OBJECT_HPP_
+#endif  // RVIZ_VISUAL_TESTS__PAGE_OBJECTS__CAMERA_DISPLAY_PAGE_OBJECT_HPP_
