+++ rviz_visual_tests/include/rviz_visual_tests/visual_test_fixture.hpp
@@ -27,36 +27,53 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef INTERNAL__IMAGE_TESTER_HPP_
-#define INTERNAL__IMAGE_TESTER_HPP_
+#ifndef RVIZ_VISUAL_TESTS__VISUAL_TEST_FIXTURE_HPP_
+#define RVIZ_VISUAL_TESTS__VISUAL_TEST_FIXTURE_HPP_
 
-#ifdef _WIN32
-# pragma warning(push)
-# pragma warning(disable : 4996)
-#endif
-#include <Ogre.h>
-#ifdef _WIN32
-# pragma warning(pop)
-#endif
+#include <memory>
+#include <vector>
 
-class ImageTester
+#include "rviz_visual_tests/internal/display_handler.hpp"
+#include "rviz_visual_tests/internal/executor.hpp"
+#include "rviz_visual_tests/internal/visual_test.hpp"
+#include "rviz_visual_tests/page_objects/page_object_with_window.hpp"
+
+class VisualTestFixture : public testing::Test
 {
 public:
-  ImageTester(Ogre::String reference_directory_path, Ogre::String test_directory_path);
-  void compare(Ogre::String image_name);
-  void computeImageDifference(
-    Ogre::Image test_image, Ogre::Image reference_image, Ogre::String image_name);
+  VisualTestFixture();
+  static void SetUpTestCase();
+  void TearDown() override;
+  static void TearDownTestCase();
 
-  double computeMseIndex(Ogre::String image_name, size_t image_width, size_t image_height);
+  void setCamPose(Ogre::Vector3 camera_pose);
+  void setCamLookAt(Ogre::Vector3 camera_look_at_vector);
+  template<typename T>
+  std::shared_ptr<T> addDisplay()
+  {
+    return display_handler_->addDisplay<T>();
+  }
+  void removeDisplay(std::shared_ptr<BasePageObject> display);
+
+  void captureMainWindow(Ogre::String image_name = "");
+  void captureRenderWindow(
+    std::shared_ptr<PageObjectWithWindow> display, Ogre::String name = "");
+
+  void assertScreenShotsIdentity();
+  void assertMainWindowIdentity(Ogre::String image_name = "");
+
+  Ogre::String test_name_;
+  std::unique_ptr<VisualTest> visual_test_;
+  std::unique_ptr<DisplayHandler> display_handler_;
+  std::shared_ptr<std::vector<int>> all_display_ids_vector_;
+  std::vector<Ogre::String> screen_shots_;
+  std::shared_ptr<Executor> executor_;
+  static QApplication * qapp_;
+  static rviz_common::VisualizerApp * visualizer_app_;
 
 private:
-  Ogre::Image loadImage(Ogre::String image_name, Ogre::String image_directory_path);
-  size_t pixelDifference(
-    Ogre::Image reference_image, Ogre::Image test_image, size_t width, size_t height);
-  void assertImageIdentity(
-    Ogre::String image_name, Ogre::Image reference_image, Ogre::Image test_image);
-
-  Ogre::String reference_directory_path_;
-  Ogre::String test_directory_path_;
+  void startApplication();
+  void setNameIfEmpty(Ogre::String & name);
 };
-#endif  // INTERNAL__IMAGE_TESTER_HPP_
+
+#endif  // RVIZ_VISUAL_TESTS__VISUAL_TEST_FIXTURE_HPP_
