+++ rviz_visual_tests/src/page_objects/point_cloud_display_page_object.cpp
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "point_cloud_display_page_object.hpp"
+#include "rviz_visual_tests/page_objects/point_cloud_display_page_object.hpp"
 
 #include <memory>
 #include <iostream>
@@ -39,7 +39,7 @@ PointCloudDisplayPageObject::PointCloudDisplayPageObject(
   int display_id,
   std::shared_ptr<Executor> executor,
   std::shared_ptr<std::vector<int>> all_displays_ids)
-: BasePageObject(display_id, 0, 3, executor, all_displays_ids)
+: BasePageObject(display_id, 0, 4, executor, all_displays_ids)
 {}
 
 void PointCloudDisplayPageObject::setSizeMeters(QString meters_size)
